Rename processors to entities
