+++ src/entities/inline_image.py
