+++ src/entities/reference_image.py
